Solución de: Financiamiento, error al guardar intereses. Proyección de ventas, error al calcular ingresos generados. Revisión de flujo de caja. Agregado StartWindow con logo de Negotium. Soporte a selección de moneda por proyecto.

+++ UCR.Negotium.DataAccess/CostoMensualData.cs
@@ -38,7 +38,7 @@ namespace UCR.Negotium.DataAccess
                     CostoMensual detallecosto = new CostoMensual();
                     detallecosto.CodCostoMensual = reader.GetInt32(0);
                     detallecosto.Mes = reader.GetString(1);
-                    detallecosto.CostoUnitario = reader.GetInt32(2);
+                    detallecosto.CostoUnitario = reader.GetDouble(2);
                     detallecosto.Cantidad = reader.GetDouble(3);
                     listaCostos.Add(detallecosto);
                 }//while
